+++ GenericServices/Internal/Decoders/DecodedDto.cs
@@ -39,6 +39,9 @@ namespace GenericServices.Internal.Decoders
             _matcher = new MethodCtorMatcher(publicConfig.NameMatcher);
             _perDtoConfig = perDtoConfig; //can be null
             LinkedEntityInfo = entityInfo;
+            if (entityInfo.EntityStyle == EntityStyles.DbQuery)
+                //If DbQuery then exit immediately as properties etc 
+                return;
 
             PropertyInfos = dtoType.GetProperties()
                 .Select(x => new DecodedDtoProperty(x,