+++ GenericServices/PublicButHidden/CrudServices.cs
@@ -5,6 +5,7 @@ using System;
 using System.Linq;
 using System.Linq.Expressions;
 using AutoMapper.QueryableExtensions;
+using GenericServices.Configuration.Internal;
 using GenericServices.Internal;
 using GenericServices.Internal.Decoders;
 using GenericServices.Internal.LinqBuilders;
@@ -64,6 +65,8 @@ namespace GenericServices.PublicButHidden
         {    
             T result;
             var entityInfo = _context.GetEntityInfoThrowExceptionIfNotThere(typeof(T));
+            if (entityInfo.EntityStyle == EntityStyles.DbQuery)
+                throw new InvalidOperationException($"The class {entityInfo.EntityType.Name} of style {entityInfo.EntityStyle} cannot be used in a Find.");
             if (entityInfo.EntityType == typeof(T))
             {
                 result = _context.Set<T>().Find(keys);
@@ -92,7 +95,7 @@ namespace GenericServices.PublicButHidden
             var entityInfo = _context.GetEntityInfoThrowExceptionIfNotThere(typeof(T));
             if (entityInfo.EntityType == typeof(T))
             {
-                result = _context.Set<T>().Where(whereExpression).SingleOrDefault();
+                result = entityInfo.GetReadableEntity<T>(_context).Where(whereExpression).SingleOrDefault();
             }
             else
             {
@@ -118,7 +121,7 @@ namespace GenericServices.PublicButHidden
             var entityInfo = _context.GetEntityInfoThrowExceptionIfNotThere(typeof(T));
             if (entityInfo.EntityType == typeof(T))
             {
-                return _context.Set<T>().AsNoTracking();
+                return entityInfo.GetReadableEntity<T>(_context).AsNoTracking();
             }
 
             //else its a DTO, so we need to project the entity to the DTO 
@@ -130,6 +133,7 @@ namespace GenericServices.PublicButHidden
         public T CreateAndSave<T>(T entityOrDto, string ctorOrStaticMethodName = null) where T : class
         {
             var entityInfo = _context.GetEntityInfoThrowExceptionIfNotThere(typeof(T));
+            entityInfo.CheckCanDoOperation(CrudTypes.Create);
             Message = $"Successfully created a {entityInfo.EntityType.GetNameForClass()}";
             if (entityInfo.EntityType == typeof(T))
             {
@@ -158,6 +162,7 @@ namespace GenericServices.PublicButHidden
         public void UpdateAndSave<T>(T entityOrDto, string methodName = null) where T : class
         {
             var entityInfo = _context.GetEntityInfoThrowExceptionIfNotThere(typeof(T));
+            entityInfo.CheckCanDoOperation(CrudTypes.Update);
             Message = $"Successfully updated the {entityInfo.EntityType.GetNameForClass()}";
             if (entityInfo.EntityType == typeof(T))
             {
@@ -202,6 +207,7 @@ namespace GenericServices.PublicButHidden
             where TEntity : class
         {
             var entityInfo = _context.GetEntityInfoThrowExceptionIfNotThere(typeof(TEntity));
+            entityInfo.CheckCanDoOperation(CrudTypes.Update);
             Message = $"Successfully updated the {entityInfo.EntityType.GetNameForClass()}";
             if (entityInfo.EntityType != typeof(TEntity))
                 throw new NotImplementedException(
@@ -224,6 +230,7 @@ namespace GenericServices.PublicButHidden
         public void DeleteAndSave<TEntity>(params object[] keys) where TEntity : class
         {
             var entityInfo = _context.GetEntityInfoThrowExceptionIfNotThere(typeof(TEntity));
+            entityInfo.CheckCanDoOperation(CrudTypes.Delete);
             Message = $"Successfully deleted a {entityInfo.EntityType.GetNameForClass()}";
             if (entityInfo.EntityType != typeof(TEntity))
                 throw new NotImplementedException(
@@ -247,6 +254,7 @@ namespace GenericServices.PublicButHidden
             params object[] keys) where TEntity : class
         {
             var entityInfo = _context.GetEntityInfoThrowExceptionIfNotThere(typeof(TEntity));
+            entityInfo.CheckCanDoOperation(CrudTypes.Delete);
             Message = $"Successfully deleted a {entityInfo.EntityType.GetNameForClass()}";
             if (entityInfo.EntityType != typeof(TEntity))
                 throw new NotImplementedException(