+++ Tests/EfClasses/Child.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) 2018 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+// Licensed under MIT license. See License.txt in the project root for license information.
+
+namespace Tests.EfClasses
+{
+    public class Child
+    {
+        public int ChildId { get; set; }
+        public string MyString { get; set; }
+    }
+}