+++ Tests/EfClasses/ParentWithChildCountQuery.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) 2018 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+// Licensed under MIT license. See License.txt in the project root for license information.
+
+namespace Tests.EfClasses
+{
+    public class ParentWithChildCountQuery
+    {
+        public int ParentId { get; set; }
+        public int NumChildren { get; set; }
+    }
+}