+++ Tests/UnitTests/GenericServicesSetup/TestSetupDtosAndMappings.cs
@@ -95,7 +95,7 @@ namespace Tests.UnitTests.GenericServicesSetup
 
             //VERIFY
             setupDtos.IsValid.ShouldBeFalse();
-            setupDtos.Errors.Count.ShouldEqual(3);    
+            setupDtos.Errors.Count.ShouldEqual(4);    
         }